fix(ci): align slice coverage mapping and package helpers
