+++ crates/ramshared-winsvc/src/main.rs
@@ -901,7 +901,7 @@ mod windows_svc {
         let config = WinDriveConfig::from_reader(&std::fs::read(
             root.join(&manifest.artifact(ArtifactRole::WinsvcConfig)?.relative_path),
         )?)?;
-        let script = build_residue_script(config.volume_letter)?;
+        let script = ramshared_winsvc::package::build_residue_script(config.volume_letter)?;
         let mut child = std::process::Command::new("powershell.exe")
             .args(["-NoProfile", "-NonInteractive", "-Command", &script])
             .stdout(std::process::Stdio::piped())
@@ -1161,42 +1161,3 @@ fn main() {
     }
 }
 
-pub fn build_residue_script(letter: char) -> Result<String, Box<dyn std::error::Error>> {
-    let uppercase_letter = letter.to_ascii_uppercase();
-    if !uppercase_letter.is_ascii_uppercase() {
-        return Err("invalid volume letter for host residue query".into());
-    }
-    let escaped_letter = uppercase_letter.to_string().replace('\'', "''");
-    Ok(format!(
-        concat!(
-            "$ErrorActionPreference='Stop';",
-            "$d=@(Get-CimInstance Win32_DiskDrive|?{{",
-            "$_.Model -match 'RAMSHARE|VRAMDISK' -or $_.Caption -match 'RAMSHARE|VRAMDISK'}});",
-            "$p=@(Get-CimInstance Win32_PageFileUsage|?{{",
-            "$_.Name -match '^{letter}:\\\\'}});",
-            "Write-Output ($d.Count.ToString()+'|'+$p.Count.ToString())"
-        ),
-        letter = escaped_letter
-    ))
-}
-
-#[cfg(test)]
-mod tests {
-    #![allow(clippy::unwrap_used, clippy::expect_used)]
-    use super::*;
-
-    #[test]
-    fn observe_host_residue_script_escaping() {
-        let valid_script = build_residue_script('D').unwrap();
-        assert!(valid_script.contains("$_.Name -match '^D:\\\\'"));
-
-        let lowercase_script = build_residue_script('z').unwrap();
-        assert!(lowercase_script.contains("$_.Name -match '^Z:\\\\'"));
-
-        let invalid_char = build_residue_script('1');
-        assert!(invalid_char.is_err());
-
-        let injected_char = build_residue_script('\'');
-        assert!(injected_char.is_err());
-    }
-}
