+++ docs/benchmarks/history/benchmark-2026-08-29_06-09-51.json
@@ -1,0 +1,20 @@
+{
+  "battery_mode": false,
+  "cascade_mode": false,
+  "max_safe_pct": 2,
+  "total_allocated_mb": 318,
+  "peak_swap_mb": 395,
+  "tier1_zram_mb": 343,
+  "tier1_zram_pct": 33,
+  "tier2_vram_mb": 28,
+  "tier2_vram_pct": 0,
+  "tier3_ssd_mb": 24,
+  "tier3_ssd_pct": 0,
+  "peak_pressure_index": 1.3557697859571374,
+  "telemetry_readings_count": 2,
+  "active_io_cycles_completed": 0,
+  "reclaim_duration_ms": 76.282084,
+  "reclaim_speed_gbs": 4.071032917768738,
+  "post_reclaim_free_ram_mb": 8488,
+  "status": "PASS_ZERO_PANIC"
+}
