+++ docs/benchmarks/history/benchmark-2026-08-29_06-09-52.json
@@ -1,0 +1,20 @@
+{
+  "battery_mode": true,
+  "cascade_mode": false,
+  "max_safe_pct": 2,
+  "total_allocated_mb": 318,
+  "peak_swap_mb": 395,
+  "tier1_zram_mb": 343,
+  "tier1_zram_pct": 33,
+  "tier2_vram_mb": 28,
+  "tier2_vram_pct": 0,
+  "tier3_ssd_mb": 24,
+  "tier3_ssd_pct": 0,
+  "peak_pressure_index": 1.389989459414275,
+  "telemetry_readings_count": 4,
+  "active_io_cycles_completed": 2,
+  "reclaim_duration_ms": 68.198919,
+  "reclaim_speed_gbs": 4.55354541616708,
+  "post_reclaim_free_ram_mb": 8650,
+  "status": "PASS_ZERO_PANIC"
+}
