+++ tmp-ci-rust-slice-changes.txt
@@ -1,0 +1,9 @@
+Cargo.lock
+crates/ramshared-agent/src/watchdog.rs
+crates/ramshared-broker/src/protocol.rs
+crates/ramshared-cuda/src/vram_impl.rs
+crates/ramshared-tier/src/cascade.rs
+crates/ramshared-tier/src/n3_state.rs
+crates/ramshared-uring/src/lib.rs
+crates/ramshared-winsvc/src/main.rs
+docs/benchmarks/history/latest.json
